- Fixed bags  - Fixed some missed new addition
git-svn-id: https://subversion.assembla.com/svn/mapler-me@484 7cbc8f23-9b70-4854-b091-be65f742de0c

+++ branches/MPLRServer/Internal Storage/Store.cs
@@ -57,7 +57,7 @@ namespace MPLRServer.Internal_Storage
             return KnownCharlist[tmp.ID][tmp.WorldID];
         }
 
-        public void SetChecksumOfSlot(int pCharacterID, byte pWorldID, byte pInventory, short pSlot, int pChecksum)
+        public void SetChecksumOfSlot(int pCharacterID, byte pWorldID, ushort pInventory, short pSlot, int pChecksum)
         {
             if (!Internal_Storage.Store.Instance.KnownCharlist[pCharacterID][pWorldID].SlotHashes.ContainsKey(pInventory))
                 Internal_Storage.Store.Instance.KnownCharlist[pCharacterID][pWorldID].SlotHashes.Add(pInventory, new Dictionary<short, int>());
@@ -147,7 +147,7 @@ namespace MPLRServer.Internal_Storage
                         }
                     }
 
-                    int inventory = result.GetInt32("inventory");
+                    ushort inventory = result.GetUInt16("inventory");
                     if (!ch.SlotHashes.ContainsKey(inventory))
                         ch.SlotHashes.Add(inventory, new Dictionary<short, int>());
